[2.4.0] AbstractArangoClusterHealthIndicator added to Facilitate LegacyArangoClusterHealthIndicator and ArangoClusterHealthIndicator logic Javadoc improved for ArangoProperties and AbstractArangoConfiguration Tests for legacy health indicator added
