--- src/main/java/io/micronaut/configuration/arango/EnableConfiguration.java
@@ -1,24 +1,0 @@
-package io.micronaut.configuration.arango;
-
-/**
- * Configuration for enabled properties
- *
- * @author Anton Kurako (GoodforGod)
- * @since 23.10.2020
- */
-public class EnableConfiguration {
-
-    private boolean enabled;
-
-    public EnableConfiguration(boolean enabled) {
-        this.enabled = enabled;
-    }
-
-    public boolean isEnabled() {
-        return enabled;
-    }
-
-    public void setEnabled(boolean enabled) {
-        this.enabled = enabled;
-    }
-}
