+++ src/main/java/io/micronaut/configuration/arango/health/ArangoClusterHealthConfiguration.java
@@ -9,7 +9,7 @@ import io.micronaut.context.annotation.Requires;
  * @since 13.08.2021
  */
 @Requires(property = ArangoSettings.PREFIX)
-@ConfigurationProperties(ArangoSettings.PREFIX + ".health-cluster")
+@ConfigurationProperties(ArangoSettings.PREFIX + ".health.cluster")
 public class ArangoClusterHealthConfiguration extends AbstractHealthConfiguration {
 
 }
