+++ src/test/java/io/micronaut/configuration/arango/ArangoRunner.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Assertions;
 public abstract class ArangoRunner extends Assertions {
 
     protected static ArangoContainer getContainer() {
-        return new ArangoContainer(ArangoContainer.LATEST)
+        return new ArangoContainer("3.7.13")
                 .withoutAuth()
                 .withFixedPort(ArangoContainer.PORT_DEFAULT);
     }
